Rough fix on ResourceWarning
Following the hard work of sigmavirus24 in [this issue](https://github.com/kennethreitz/requests/issues/1685#issuecomment-62152377)
I added a function that just closes connections after every moesearch request.

This is a rough fix: if you're using moesearch for "something serious", you should wait
until a new version of Requests and wait for it to be fixed the proper way.